requirements.txt
feedparser==6.0.11
python-dotenv==1.0.1
google-api-python-client==2.118.0
firecrawl-py==0.0.16
youtube-transcript-api==0.6.2
google-generativeai==0.4.0
python-dateutil==2.8.2
